move check_new_nodes to utils
